requirements.txt
flask
requests
openai
python-dotenv
gunicorn
flask-restful